% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscon.R
\name{mscon}
\alias{mscon}
\title{Connect to mssql server}
\usage{
mscon(dsn, database, path, DBMSencoding = "UTF-8")
}
\arguments{
\item{dsn}{is a data source name}

\item{database}{is a database}

\item{path}{is a search path}

\item{DBMSencoding}{is a dbms encoding}
}
\description{
Connect to mssql server
}
\keyword{connector}
\keyword{ms-sql}
