% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msconn.R
\name{msconn}
\alias{msconn}
\title{Connect to mssql server}
\usage{
msconn(dsn, db, path)
}
\arguments{
\item{dsn}{is a data source name}

\item{path}{is a search path}

\item{database}{is a database}
}
\description{
Connect to mssql server
}
\keyword{connector}
\keyword{ms-sql}
