% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pycon.R
\name{pycon}
\alias{pycon}
\title{Connect to mssql server with python}
\usage{
pycon(condaenv, server, database, charset = "UTF-8")
}
\arguments{
\item{condaenv}{is a anaconda virtual environment name}

\item{server}{is a server}

\item{database}{is a database}

\item{charset}{is a character set}
}
\description{
Connect to mssql server with python
}
\keyword{connector}
\keyword{ms-sql}
\keyword{python}
\keyword{with}
